feat(proxy wrapper): replace revocable proxies with lockable proxies
previously, there was a revoke method on the proxy wrapper which served as a way to permanently
revoke a proxy. That method is now removed and has been replaced with a custom mechanism to lock and
unnlock proxies. This provides the same functionality as revoking but has the advantage that can be
undone.